analytics/requirements.txt
affine==2.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contextily==1.7.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mercantile==1.2.1
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pyarrow==22.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
scipy==1.16.3
shapely==2.1.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.11.0
